# Stop copying LogicalPlan and Exprs in `OptimizeProjections` (2% faster planning)