… datafusion/optimizer/src/optimize_projections/mod.rs